+++ barramento/src/main/java/br/anhembi/barramento/repo/repository.java
@@ -2,7 +2,7 @@ package br.anhembi.barramento.repo;
 import br.anhembi.barramento.model.model;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface repository extends JpaRepository<model, Integer> {
-    boolean existsByUserIdAndData(int userId, java.time.LocalDate data);
+public interface repository extends JpaRepository<model, Long> {
+    boolean existsByUserIdAndData(Long userId, java.time.LocalDate data);
 }
